+++ pypeit/spectrographs/vlt_xshooter.py
@@ -344,7 +344,11 @@ class VLTXShooterNIRSpectrograph(VLTXShooterSpectrograph):
         # Sensitivity function parameters
         par['sensfunc']['algorithm'] = 'IR'
         par['sensfunc']['polyorder'] = 8
-        par['sensfunc']['IR']['telgridfile'] = 'TelFit_Paranal_NIR_9800_25000_R25000.fits'
+        par['sensfunc']['IR']['telgridfile'] = 'TellPCA_3000_26000_R25000.fits'
+        par['sensfunc']['IR']['pix_shift_bounds'] = (-8.0,8.0)
+        
+        # Telluric parameters
+        par['telluric']['pix_shift_bounds'] = (-8.0,8.0)
 
         # Coadding
         par['coadd1d']['wave_method'] = 'log10'
@@ -720,7 +724,11 @@ class VLTXShooterVISSpectrograph(VLTXShooterSpectrograph):
         # Sensitivity function parameters
         par['sensfunc']['algorithm'] = 'IR'
         par['sensfunc']['polyorder'] = 8 #[9, 11, 11, 9, 9, 8, 8, 7, 7, 7, 7, 7, 7, 7, 7]
-        par['sensfunc']['IR']['telgridfile'] = 'TelFit_Paranal_VIS_4900_11100_R25000.fits'
+        par['sensfunc']['IR']['telgridfile'] = 'TellPCA_3000_26000_R25000.fits'
+        par['sensfunc']['IR']['pix_shift_bounds'] = (-8.0,8.0)
+        
+        # Telluric parameters
+        par['telluric']['pix_shift_bounds'] = (-8.0,8.0)
 
         # Coadding
         par['coadd1d']['wave_method'] = 'log10'
@@ -1009,8 +1017,11 @@ class VLTXShooterUVBSpectrograph(VLTXShooterSpectrograph):
         # Sensitivity function parameters
         par['sensfunc']['algorithm'] = 'IR'
         par['sensfunc']['polyorder'] =  8
-        par['sensfunc']['IR']['telgridfile'] = 'TelFit_LasCampanas_3100_26100_R20000.fits'
-        # This is a hack until we we have a Paranal file generated that covers the UVB wavelength range.
+        par['sensfunc']['IR']['telgridfile'] = 'TellPCA_3000_26000_R25000.fits'
+        par['sensfunc']['IR']['pix_shift_bounds'] = (-8.0,8.0)
+        
+        # Telluric parameters
+        par['telluric']['pix_shift_bounds'] = (-8.0,8.0)
 
         # Coadding
         par['coadd1d']['wave_method'] = 'log10'